git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@8870 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/uFldCollisionDetect/CollisionDetector.cpp
@@ -307,9 +307,9 @@ bool CollisionDetector::OnStartUp()
   m_param_summary += ", encounter_range=" + doubleToStringX(m_encounter_dist);
   Notify("COLLISION_DETECT_PARAMS", m_param_summary);
 
-  Notify("ENCOUNTER_TOTAL", 0);
-  Notify("NEAR_MISS_TOTAL", 0);
-  Notify("COLLISION_TOTAL", 0);
+  Notify("ENCOUNTER_TOTAL", (double)0);
+  Notify("NEAR_MISS_TOTAL", (double)0);
+  Notify("COLLISION_TOTAL", (double)0);
   
   registerVariables();	
   return(true);
